+++ src/main/kotlin/dev/panini/sankhya/SankhyaGenerator.kt
@@ -1,7 +1,6 @@
 package dev.panini.sankhya
 
 import dev.panini.derivation.DerivationResult
-import java.math.BigInteger
 
 /** Public numeral API. Every method returns an auditable grammatical derivation. */
 class SankhyaGenerator(
@@ -9,14 +8,14 @@ class SankhyaGenerator(
     private val puranaDeriver: PuranaSankhyaDeriver = PuranaSankhyaDeriver(cardinalDeriver),
 ) {
     /** Numeric identity of a canonical annotated prātipadika; surface forms are not accepted. */
-    fun annotatedPratipadikaValue(pratipadika: String): BigInteger? =
+    fun annotatedPratipadikaValue(pratipadika: String): Long? =
         PrimitiveSankhya.fromAnnotatedPratipadika(pratipadika)?.value
 
-    fun cardinal(value: BigInteger): DerivationResult = cardinalDeriver.derive(value)
+    fun cardinal(value: Long): DerivationResult = cardinalDeriver.derive(value)
 
-    fun cardinalVariants(value: BigInteger): List<DerivationResult> = cardinalDeriver.deriveVariants(value)
+    fun cardinalVariants(value: Long): List<DerivationResult> = cardinalDeriver.deriveVariants(value)
 
-    fun ordinal(value: BigInteger): DerivationResult = puranaDeriver.derive(value)
+    fun ordinal(value: Long): DerivationResult = puranaDeriver.derive(value)
 
-    fun ordinalVariants(value: BigInteger): List<DerivationResult> = puranaDeriver.deriveVariants(value)
+    fun ordinalVariants(value: Long): List<DerivationResult> = puranaDeriver.deriveVariants(value)
 }
